requirements.txt
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==2.2.3
zipp==3.15.0
